Prevent password change email getting sent several times
This is in there as a safeguard, if you save a user model with a new
.password, it'll update it, but it won't remove it so that you still
have access to that password, that, however, means that if you update
again the same email will be sent because the .password is in there.

So I setup a safeguard so that doesn't happen.