Seperate all the tests for test running to be more modular
Now you can run unit tests (with test:unit), integration tests for the
InstallationManager (with test:integration:manager) and the integration
tests for the installations (with test:integration:installation).

This doesn't include the actual fixed tests, this commit only contains
the new scripts that allow you to run stuff individually.

The test/integration-manager.js file is not there yet though, because I
haven't gotten to that part of the system.